+++ plotnine/positions/position.py
@@ -3,6 +3,7 @@ from warnings import warn
 
 import numpy as np
 
+from ..mapping.aes import X_AESTHETICS, Y_AESTHETICS
 from ..utils import check_required_aesthetics, groupby_apply
 from ..utils import is_string, Registry
 from ..exceptions import PlotnineError, PlotnineWarning
@@ -88,16 +89,12 @@ class position(metaclass=Registry):
             Transforms y scale mappings
             Takes one argument, either a scalar or an array-type
         """
-        # Aesthetics that map onto the x and y scales
-        X = {'x', 'xmin', 'xmax', 'xend', 'xintercept'}
-        Y = {'y', 'ymin', 'ymax', 'yend', 'yintercept'}
-
         if trans_x:
-            xs = [name for name in data.columns if name in X]
+            xs = [name for name in data.columns if name in X_AESTHETICS]
             data[xs] = data[xs].apply(trans_x)
 
         if trans_y:
-            ys = [name for name in data.columns if name in Y]
+            ys = [name for name in data.columns if name in Y_AESTHETICS]
             data[ys] = data[ys].apply(trans_y)
 
         return data